requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
gunicorn==21.2.0
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
whitenoise==6.6.0
django-storages==1.14.2
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
bandit==1.7.5
safety==2.3.5
black==23.9.1
isort==5.12.0
flake8==6.1.0 